Extract hex-grid crate: layout + math with SVG example
Move HGridLayout and pure math functions (edge_cuboid_transform,
gap_filler, gap_vertex_data, idw_interpolate_height, map_noise_to_range)
into a standalone crate with zero Bevy/ECS dependencies — only glam,
hexx, and noise. h_terrain now re-exports through thin wrapper modules.

Includes an SVG example (cargo run -p hex-grid --example svg) that
renders a top-down 2D hex grid colored by terrain height.

78 tests pass (28 in hex-grid, 50 in hex-terrain).
