+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/mesh-gradient"]
+members = ["crates/mesh-gradient", "crates/hex-grid"]
 
 [package]
 name = "hex-terrain"
@@ -52,6 +52,7 @@ bevy-inspector-egui = "0.36"
 bevy_egui = "0.39"
 hexx = { version = "0.24.0", features = ["bevy"] }
 mesh-gradient = { path = "crates/mesh-gradient" }
+hex-grid = { path = "crates/hex-grid" }
 
 [lints.rust]
 unused = { level = "deny" }
